+++ contrib/runners/mistral_v2/tests/unit/test_mistral_v2_cancel.py
@@ -71,8 +71,8 @@ WF1 = workflows.Workflow(None, {'name': WF1_NAME, 'definition': WF1_YAML})
 WF1_OLD = workflows.Workflow(None, {'name': WF1_NAME, 'definition': ''})
 WF1_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
 WF1_EXEC['workflow_name'] = WF1_NAME
-WF1_EXEC_PAUSED = copy.deepcopy(WF1_EXEC)
-WF1_EXEC_PAUSED['state'] = 'PAUSED'
+WF1_EXEC_CANCELLED = copy.deepcopy(WF1_EXEC)
+WF1_EXEC_CANCELLED['state'] = 'CANCELLED'
 
 
 @mock.patch.object(
@@ -130,7 +130,7 @@ class MistralRunnerCancelTest(DbTestCase):
         mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
     @mock.patch.object(
         executions.ExecutionManager, 'update',
-        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC_PAUSED)))
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC_CANCELLED)))
     def test_cancel(self):
         liveaction = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
         liveaction, execution = action_service.request(liveaction)
@@ -144,7 +144,7 @@ class MistralRunnerCancelTest(DbTestCase):
 
         requester = cfg.CONF.system_user.user
         liveaction, execution = action_service.request_cancellation(liveaction, requester)
-        executions.ExecutionManager.update.assert_called_with(WF1_EXEC.get('id'), 'PAUSED')
+        executions.ExecutionManager.update.assert_called_with(WF1_EXEC.get('id'), 'CANCELLED')
         liveaction = LiveAction.get_by_id(str(liveaction.id))
         self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_CANCELED)
 
@@ -163,7 +163,7 @@ class MistralRunnerCancelTest(DbTestCase):
     @mock.patch.object(
         executions.ExecutionManager, 'update',
         mock.MagicMock(side_effect=[requests.exceptions.ConnectionError(),
-                                    executions.Execution(None, WF1_EXEC_PAUSED)]))
+                                    executions.Execution(None, WF1_EXEC_CANCELLED)]))
     def test_cancel_retry(self):
         liveaction = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
         liveaction, execution = action_service.request(liveaction)
@@ -177,7 +177,7 @@ class MistralRunnerCancelTest(DbTestCase):
 
         requester = cfg.CONF.system_user.user
         liveaction, execution = action_service.request_cancellation(liveaction, requester)
-        executions.ExecutionManager.update.assert_called_with(WF1_EXEC.get('id'), 'PAUSED')
+        executions.ExecutionManager.update.assert_called_with(WF1_EXEC.get('id'), 'CANCELLED')
         liveaction = LiveAction.get_by_id(str(liveaction.id))
         self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_CANCELED)
 
@@ -210,7 +210,7 @@ class MistralRunnerCancelTest(DbTestCase):
         requester = cfg.CONF.system_user.user
         liveaction, execution = action_service.request_cancellation(liveaction, requester)
 
-        calls = [call(WF1_EXEC.get('id'), 'PAUSED') for i in range(0, 2)]
+        calls = [call(WF1_EXEC.get('id'), 'CANCELLED') for i in range(0, 2)]
         executions.ExecutionManager.update.assert_has_calls(calls)
 
         liveaction = LiveAction.get_by_id(str(liveaction.id))